+++ ps2host/net_fsys.c
@@ -230,7 +230,7 @@ static int fsysLseek(int fd, unsigned int offset, int whence)
     return ret;
 }
 //----------------------------------------------------------------------------
-static int fsysIoctl(iop_file_t *file, unsigned long request, void *data)
+static int fsysIoctl(iop_file_t *file, int request, void *data)
 {
     int ret;
     dbgprintf("fsysioctl..\n");
